+++ src/nodes/animation/create.ts
@@ -66,7 +66,7 @@ class AnimationNode extends TriggerNode<
 		if (!("kind" in loc))
 			return { x: loc.x, y: loc.y };
 
-		switch (loc.kind) {
+		switch (loc?.kind) {
 			case "point":
 				return { x: loc.x, y: loc.y };
 			case "region":