+++ src/nodes/canvasPan/base.ts
@@ -64,7 +64,7 @@ abstract class CanvasPanModifierNode<
 		if (!("kind" in loc))
 			return { x: loc.x, y: loc.y };
 
-		switch (loc.kind) {
+		switch (loc?.kind) {
 			case "point":
 				return { x: loc.x, y: loc.y };
 			case "region":