+++ src/nodes/canvasPan/create.ts
@@ -113,7 +113,7 @@ class CanvasPanNode extends TriggerNode<
 		if (!("kind" in loc))
 			return { x: loc.x, y: loc.y };
 
-		switch (loc.kind) {
+		switch (loc?.kind) {
 			case "point":
 				return { x: loc.x, y: loc.y };
 			case "region":