+++ src/nodes/crosshair/base.ts
@@ -59,8 +59,8 @@ abstract class CrosshairModifierNode<
 
 	protected abstract apply(section: CrosshairSection): Promise<void> | void;
 
-	getLocation(loc: PositionSource): TokenDocument | Point | RegionDocument | string | undefined {
-		switch (loc.kind) {
+	getLocation(loc: PositionSource | undefined): TokenDocument | Point | RegionDocument | string | undefined {
+		switch (loc?.kind) {
 			case "point":
 				return { x: loc.x, y: loc.y };
 			case "region":