+++ src/nodes/effect/base.ts
@@ -105,7 +105,7 @@ abstract class EffectModifierNode<
 		if (!("kind" in loc))
 			return { x: loc.x, y: loc.y };
 
-		switch (loc.kind) {
+		switch (loc?.kind) {
 			case "point":
 				return { x: loc.x, y: loc.y };
 			case "region":