+++ src/nodes/scrollingText/base.ts
@@ -57,8 +57,8 @@ abstract class ScrollingTextModifierNode<
 		return null;
 	}
 
-	getLocation(loc: PositionSource): TokenDocument | Point | RegionDocument | string | undefined {
-		switch (loc.kind) {
+	getLocation(loc: PositionSource | undefined): TokenDocument | Point | RegionDocument | string | undefined {
+		switch (loc?.kind) {
 			case "point":
 				return { x: loc.x, y: loc.y };
 			case "region":