+++ src/nodes/utility/namedLocation.ts
@@ -70,8 +70,8 @@ class NamedLocationNode extends TriggerNode<
 		return this.executeNext("out");
 	}
 
-	getLocation(loc: PositionSource): TokenDocument | Point | RegionDocument | string | undefined {
-		switch (loc.kind) {
+	getLocation(loc: PositionSource | undefined): TokenDocument | Point | RegionDocument | string | undefined {
+		switch (loc?.kind) {
 			case "point":
 				return { x: loc.x, y: loc.y };
 			case "region":